Working hasFree and nextFree in Transaction Table
After making the connection that the syntax `X => Y` means that Scala
wants an input that is a function mapping `X` to `Y`, the whole
structure of the Vec class makes sense. All those methods (like
`exists` and `indexWhere`) just need a function that takes an instance
of the Vector element and transforms it into a Bool. Thereby, I was
able to quickly define an `isFree` function and use this to derive
`hasFree` and `nextFree` values. This all appears to be working and
forms the first step of transaction entry.