Added SRAMElement Module (close #21)
New SRAMElement module that adds a wrapper around the SRAM module (the
Chisel equivalent of ram-infer) that enables sub-word writes, but
whole word reads. This is untested for dataWidths larger than 32 due
to the increasing complexity of writing a testbench that uses Scala
BigInts.

This commit additionally cleans up the SRAM module by adding a proper
sequential read on output data.