Added RRArbiter to TransactionTable
I've added a round robing arbiter (`RRArbiter` class from
ChiselUtil.scala) that selects one of the Transaction Table entries with
work to do and sends this, using a decoupled output interface, as an
output of the Transaction Table. It is intended that this output will
then be fed to DANA's control module for actual processing.