New DanaBundle Abstract Class
In the same vein as the previous commit moves everything to a DanaModule
abstract base class, this moves all Bundles to a DanaBundle abstract
base class. This allows me to define all the parameters twice (once for
DanaModule and once for DanaBundle) and all the Chisel describing
modules and interfaces can then be severely abbreviated (this deletes
some 300 lines!). The source Chisel then becomes much more readable as I
don't have to explicitly pass parameters all over the place.