Move TransactionTable Test Functions to DanaTester
The functions used to poke the TransactionTable have utility to more
than just the TTable. Specifically, a forthcoming Dana module needs to
be able to send write requests to the TTable. I've rolled in the
`newWriteRequest` and `writeRndData` functions originally defined in the
TTable into the DanaTester abstract class defined in Common.scala. These
functions now require the user to pass in a TransactionTable object (so
the functions know what to poke).