New Module-generic Info Function in DanaTester
I've added a new module-agnostic function called `info` to the
DanaTester class (in Common.scala). Based on the type that this function
is passed, it will print out Module-specific information about that
module. The intent here is that this function will hold _all information
related to printing the state of all modules in the dana package_. That
way, any module that instantiates an instance of another module will
be able to print information about it.