Add WIP DanaControl Module, Constants Definition
I've added a new WIP control module for DANA. This lives in
DanaControl.scala. Currently, this just responds to one basic type of
request generated by the Transaction Table.

A top-level module in Dana.scala is being used to instantiate and wire
everything up.

As these modules are passing back and forth some shared state
information, I needed to define a good way to provide for shared
enumerated types. However, Chisel doesn't seem to support this, so I'm
using a `trait` of `Constants` defined in Constants.scala to declare
shared state types. This is kludgy, but it works. As support for this,
I've included a Packages.scala in the same way that the Rocket core uses
a packages.scala file.