Added unused half-period clock to Dana.scala
In an initial attempt to get a clock waveform to show up in the VCD
file, I've added hack from the Chisel Users Group that creates a clock
running at twice the frequency of the default clock. However, this still
needs to be manually dumped to the VCD file from the C++ testbench. I
don't have a good workaround for doing this currently.

The Makefile also has some unused targets related to this. Specifically,
when using additional clock domains the generated XXX-emulator.cpp
defines a `set_clocks` (name may be wrong). This needs to be included
with any subsequent testbenches. So, I went ahead and added an
target/rule that will strip out the main symbol from XXX-emulator.o to
create XXX-emulator-nomain.o. This can then be compiled to build the
actual testbench that you care about.