First Layer Update Working for DANA-Chisel
The Chisel version of DANA is now handling the first cache layer
requests from the Transaction Table and the Control module. Note, that
this differs somewhat from the SystemVerilog version (nnsim-hdl) in that
the addresses are communicated directly to the cache SRAMs and are not
passed through an asynchronous interface to individual cache entry
modules. This may result in timing problems later on (as this removes a
pipeline stage).

Also, the wicked kludge of using the "retry" counters in the
SystemVerilog Control module has been (correctly) abandoned. An
additional field, generically called "waiting", has been defined in the
TransactionTable that indicates if a Transaction is waiting for
something before it can proceed. This has similar functionality in
indicating that a specific TransactionTable entry has work to do and is
used to indicate to the TransactionTable round robin arbiter that it is
"valid". This replaces the waitingForCache bit used in the SystemVerilog
version and in earlier Chisel variants.