Added Pseudo-Typecasting in Cache
As unions are not presently supported in Chisel, it becomes very
difficult to deal with data, like that coming out of the cache, which
needs to be interpreted in one of several different ways. One block of
data in the cache can describe general NN configuration info, layers,
neurons, or weights. The cache needs to be able to interpret this output
and pull out only the specific data that it cares about. If the indices
are fixed, this is trivial. however, if the indices have a variable
offset (i.e., a mux should be used), Chisel shits the bed. What is
really needed here is a way to interpret arbitrary data (i.e., a UInt)
as a bundle or a vector.

To get around this, I create explicit instances for each of these
possibly interpreted formats. The logic is then hard-coded to use a
different interpretation at run time. _This is ridiculously kludgy_, but
it does work. This does result in an increased number of inferred Chisel
nodes, however, as these are redundant representations, intelligent
synthesis tools should rip them out. This commit includes the changes
necessary to implement this.

An alternative approach would be to define a new type extending UInt
with methods to interpret its underlying data in different ways. This
seems to be the best approach (as it should result in no Chisel node
overhead), but currently isn't used.