WIP Commit re: PE Table Support
The PE Table is getting an initial kick from the Control module to
allocate a PE. The testbench t_Dana.cpp has been updated to dump all the
fields from the PE Table. Something is broken with the TID and the PE
Table seems to be allocating in reverse order (something which I thought
the built in VecLike methods didn't do). Anyway, WIP commit which should
get squashed at some point.