Minor fixes to IO location logic
The IO locations for registers were technically reversed from the most
basic case. This doesn't matter, but it's more logical now---the first
location is used first, and the second location is used second.

Input/Weight Location and Index were being assigned to each other. This
is now fixed.

The C++ testbench was mixing up Input and Weight Valid signals.