WIP -- PE Kicking for Valid Inputs/Weights Working
Using some kludgy logic split across two when blocks, I'm kicking the PE
from the PE Table when valid inputs/weights show up. The kludginess
comes from the fact that the inputs and weights can show up in the same
cycle, but the same kick logic is driven from two when blocks. I think
this is legal, but it's unsynthesizable in Verilog to drive the same
signal from two different always blocks. So, this may be a source of
error in the future.