Fix possible bug, Reg File now 1-port SRAMElement
I was not assigning the tableIndex received from a cache response in the
Control module when generating an update response to the Transaction
Table. This is now fixed.

I've changed the Register File module to use a single-port SRAMElement
variant (that is 1 read and 1 write port). I get concerned using a
2-port SRAMElement as this implies 2 read ports and 2 write ports in the
Chisel description. This should technically be okay, but if I'm going to
go that route, SRAMElement should be rewritten to use 2 Read/Write
ports.