TTable now Blocks on invalid Register File
This serves as an inefficient fix of the problem related to making sure
that Processing Elements only read valid data from the Register File.
This commit throws the Transaction Table into a waiting state when it
jumps to a new, non-output layer. The Transaction Table then waits until
the Register File indicates that it has all valid outputs. The necessary
data to perform these computations is communicated to the Register File
when a layer info request is passing through the Control module.
Pertinent information includes the total number of expected writes (the
number of neurons in the current layer) and the output location (first
or second half of the SRAM partition).

This can be greatly improved if the determination on when data is valid
is moved into the Register File. Initial attempts at providing this
functionality hit a dead end.