Add non-working SRAMElementCounter.scala
The committed module was a first pass at trying to implement the second
solution to Issue #38 before I abandonded this due to complexity.
Nevertheless, this is probably what a solution to #38 is going to look
like. This can either be modified outright, used as a base, or discarded
in favor of a better approach.