Possibly working DANA-Chisel
The Chisel variant of DANA is working for a basic testbench. I'm saying
"correct" here because I'm not sure what the exact output should be,
however, this matches the internal computation. In other words, outputs
are being generated correctly though the correctness of outputs is
untested.

This commit begins the testing/closeout for DANA-Chisel.