Fix DANA-Chisel Address Bugs
There was some latent problems related to using the correct address in
the Transaction Table. Specifically, this should standardize on the
TTable always seeing element addresses. The TTable then passes these
whole to an instance of SRAMElement and it pulls out the correct bits.
This was previously implemented inconsistently and was causing bugs (and
general confusion).