Minor Cleanup to t_Dana.cpp
Removed some commented out regions from t_Dana.cpp which are no longer
needed... and this is an excuse to close GitHub issues from a specific
commit.

Closes #17, closes #2 (we're using Chisel interfaces/Bundles and the
whole), and closes #4 (as we don't need separate software and HDL
simulators anymore).

Extensive functional testing (ideally through a regression
framework---see #1) in C++ remains to be completed as well as testing of
the Verilog that we're getting from Chisel.