gtkwave now uses head of VCD file to gen .gtkw
Updates to dana-chisel/scripts/gtkwave in light of the fact that VCD
files can be very large. GTKWave needs to be run twice: once with the
addWavesRecursive TCL script to generate a saved configuration of the
module hierarchy (.gtkw file) and a second time to actually open the
waveform. This changes that to a three step process where a truncated
version of the VCD file is used to generate the .gtkw, the VCD file is
converted to an FST file (binary file that is much smaller than a VCD
file), and GTKWave is started referencing the FST file and the saved
.gtkw. This should improve the speed with which the whole process
completes.