Change non Pow 2 parameters to division round up
Some DANA parameters can reasonably be non powers of 2. I've tried to
preemptively fix problems related to these by changing computations
related to these parameters to use a division rounded up (a Scala
function scattered in modules that need it called `divUp`). A better
solution here would probably be to just enforce parameters that are
powers of 2 (as that's what will be inferred if you use a non power of
2).