Move to X-FILES/DANA Organizational Structure
This reorganizes the original top module, Dana, into a more realistic
structure reflecting the actual organization outlined in the PACT
submission. The TransactionTable has been moved out of Dana and into a
new module called XFilesArbiter. The combined X-FILES/DANA system is
encapsulated in a new top module, Top, which instantiates an instance of
XFilesArbiter and Dana.

This necessitates minor changes to names in the top level
testbench (originally t_Dana.cpp, now t_Top.cpp) and in the Makefile as
the names of build targets have changed. Additionally, the Scala main
target is now called Top and not Dana.