DANA-Chisel Cache Properly Decrements inUseCount
When a TransactionTable entry trips to a "done" state (assertion of the
`done` flag), this will now generate a request (that has to get through
the TransactionTable round robin arbiter) to the Cache via the Control
module to decrement the `inUseCount` governing the number of outstanding
transactions that are using a specific cache entry.

This also includes an assertion to check that no `inUseCount` decrement
requests are received when the `inUseCount` is zero.