Separate Cache and PETable Ready Signals and Logic
Previously, the ready signal passed to the Transaction Table was
dependent on the Cache _and_ the Processing Element Table being able to
accept requests. This was too stringent of a requirement and this
resulted in requests from the Transaction Table which should not be
blocked getting delayed until all Processing Elements became free. To
circumvent this, I've added explicit signals indicating the readiness of
the Cache and the PE Table. These are passed, unconventionally, via the
Control module to TTable response bundle.

I've updated the assertions to reflect these changes and added another
assertion that prevents the Transaction Table from allocating
non-existent PEs beyond the limit of the network size.

The combined X-FILES/DANA system is now passing test cases involving 2
TTable entries with 1--8 PEs and 4 elements per block.