Assertion fix and general cleanup
This corrects the broad assertion in the Transaction Table on multiple
inbound requests/responses on the same transaction. The split to
separate Cache and Register File valid signals requires additional
checking which this implements. I also went ahead and checked all
existing assertions to get them into a readable format (i.e.
"![condition]" as opposed to "[!condition]").

This covers general cleanup of the top-level testbench. This now
includes support for specifying the error bound used for the error
check. All outputs are now checked for bit flips regardless of whether
or not they violate the error bound.

This adds a quick fix for the cycle count testbench parameter. The cycle
count was previously cumulative. This was not my original intention.