Remove unused signals from PE Table
The PE Table originally had 3 unused signals wrapped in debug
definitions to facilitate debugging: ASID, TID, and Node Index. The
first two are unnecessary from the perspective of the PE Table as this
doesn't care _who_ owns what just that it knows where to read and write
data and to which transaction index to respond. However knowing the
ASID/TID are useful for debugging as the user can view who owns what PE
as opposed to having to mentally dereference indices. I've removed these
from the PE Table, but have added some statements to still populate
these entries in the testbench (by using the transaction index to get
the ASID/TID from the Transaction Table). The Node Index is redundant on
the output index, so this has been outright removed.