Add reset values where needed in X-FILES/DANA
Missing reset signals for state variables (that need reset values) have
been added. This fixes bugs related to running the C++ backend with a
randomized initial state. The C++ testbench now defaults to using
randomized initial state---previously, it unknowingly did not. I've
tested this, quickly, for one set of parameters.

This also includes a WIP testbench for the Verilog backend.