X-FILES/DANA Verilog Backend OK w/ Preloaded Cache
This adds support for running a basic input/output test of the Verilog
backend of X-FILES/DANA. This includes a new testbench, t_Top.v, and two
new build targets:
* run-verilog -- Run the t_Top.v testbench
* vcd-verilog -- Run test, generate VCD, and open GTKWave

This also adds necessary support for a preloaded cache (like with
nnsim-hdl) to the X-FILES/DANA Verilog backend. Itemized changes are as
follows:
* New Verilog wrappers that connect SRAM.scala IO ports to ram_infer.v
  or ram_infer_preloaded_cache.v. These wrappers are sram.v,
  sram_infer_preloaded_cache.v, and sram_r1_w1_rw0.v.
* Modifications to SRAM.scala and SRAMElement.scala to set Verilog
  module parameters and the names of backend Verilog modules to use.
* Minor modifications to ram_infer_preloaded_cache.v due to syntax
  requirements of Icarus Verilog.
* A new configuration, DefaultFPGAConfig, in Configs.scala that sets a
  flag, PreloadCache, indicating that the cache should be preloaded.
* Modifications to the Cache to use different backend Verilog modules
  if the PreloadCache flag is asserted. Similarly, the Cache table is
  initialized on reset to valid entries when this flag is set.
* Modifications to the top level Makefile to use the DefaultFPGAConfig
  configuration when dealing with the Verilog backend.

This was tested for a cache loaded with the default values and using
Jonathan's danaCache script to populate a new cache. The build flow
expects custom caches to be located in build/cache, but this can be
changed by including the location of the cache on the Verilog search
path for Icarus Verilog.