X-FILES/DANA now Maven-like, t_Top.cpp bug fix
As I've started to pollute the X-FILES/DANA src directory with all sorts
of random files of various types (SystemVerilog, Chisel/Scala, C++),
I've seen the benefit of using a Maven-like directory structure and have
switched to this. Files are now organized primarily by function ("main"
or "test") and secondarily by language.

The C++ testbench, t_Top.cpp, was not running the single processing
tests. I changed this and modified the way that it checks if a
transaction is done to use the `is_done` method instead of `any_done`.
The latter appears to be broken.