Enable Rocket C++ Sim (breaks non-rocket backends)
This makes the necessary changes to do C++ backend simulation with a
Rocket chip. The intent is that a symlink to the "dana-chisel"
sub-directory, called "xfiles-dana", is created inside of an existing
rocket-chip repo. The `install-symlinks` script can then be run to put
the necessary files that the rocket-chip repo needs. Note: the
rocket-chip repo requires some additional modifications (which are
currently uncommitted). The needed symlinks are:
* config/XFilesDanaConfigs.scala <- X-FILES/DANA Chisel config
* src/test/cpp/xfiles_dana.h     <- These two define C++ simulation
* src/test/cpp/xfiles_dana.cc       setup like cache loading

This additionally introduces some HDL modifications to the Chisel
sources. Specifically, the top level X-FILES/DANA module now extends
class "RoCC", as required by rocket-chip. I've added support for passing
back the "rd" line when generating a response to the core. Additionally,
a read request on an unfinished ASID/TID will now generate a response of
"e_NOT_DONE" type.

The Makefile has been updated to build these RISC-V tests to
executables (with a ".rv" suffix) and disassembled versions (with a
".rvS" suffix) which will show up in the top level "build" directory.

NOTE, THE BASE CLASS CHANGE TO "RoCC" IMPLICITLY BREAKS ALL NON-ROCKET
C++/VERILOG BACKENDS BECAUSE X-FILES/DANA NOW DEPEND ON EVERYTHING IN
THE ROCKET-CHIP, ROCKET, AND UNCORE REPOS. I DON'T HAVE A GOOD SOLUTION
TO THIS, JUST YET.