Add nn-hardware tools, FANN and update build flow
This includes all the tools from nn-hardware (not as a submodule, but
as copied instances). I'm not intending to maintain nn-hardware and
this repository (xfiles-dana) should be standalone. This repository
now includes FANN as a submodule.

The build flow has been updated to support the new dependencies that
this introduces.