Update Cache to Overwrite Unused Cache Entries
Untested, but adds support for the Cache to overwrite an unused
entry (defined as having `inUseCount === 0`) when the Cache receives a
new request but doesn't have any free entries (defined as being invalid,
i.e., `valid === 0`). The erroneous case when a request shows up and
there aren't any unused or free entries is currently caught with an
assertion. However, this erroneous case should not occur as long as the
number of Transaction Table entries is less than or equal to the number
of Configuration Cache entries (since one transaction can only use one
cache entry).