Convert Register File to "Intermediate Storage"
Closes #4

This changes the storage architecture used by X-Files/Dana. The Register
File is now the only storage location removing the earlier split where
IO was stored in per-transaction SRAMs of the Transaction Table and
intermediate IO (for hidden layers) was stored in the Register File. The
Register File is now a collection of SRAMElement blocks (one per
Transaction Table entry). The underlying memory space of these SRAMs if
flat, i.e., it is up to the Transaction Table and Control modules to
determine the locations where data should be read from and written to.

Additionally, no inputs/outputs or intermediate values are ever
overwritten, as was the case in the previous implementation. While the
storage of these values increases the storage overhead of a transaction
on X-Files/Dana, though this is O(#nodes) which isn't bad, this paves
the way for error backpropagation support. Error backpropagation
requires access to all intermediate computations of the network as well
as additional storage to accumulate errors and weight updates. The flat
storage architecture of the new Register File will allow memory
addresses following the inputs, outputs, and intermediate storage of the
Register File to be used as scratch space for backprop computations.

This commit attempts to cleanup unused signals resulting from this new
storage architecture.