Add atanh-lut tool to print LUT values for atanh
This includes a new tool, tools/scripts/atanh-lut, which generates
piecewise linear x, y, and slope points similar to the existing,
hard-coded values in ActivationFunction.scala. This tool works by
generating two linear spaces, one for the original (highly sampled)
atanh, and one sub-sampled version (with a parameterized number of
knots/points). The points of the sub-sampled version are moved by a
small amount, epsilon, in the direction of an MSE minimal solution.
Epsilon starts large and gradually decreases.

This is the first step towards #5.