Fix limit problem for atanh-lut
This fixes some problems with the limits of the 32-bit numbers generated
by atanh-lut. This will now, conservatively, use one less bit of
precision if the maximum value is found to be right on the limit of the
size of a number that can be represented. This covers the case when
trying to represent (-1.0, 1.0) but this doesn't align with two's
complement (two's complement can't represent 1.0). This additionally
removes one extra bit because Java is dumb and doesn't understand
unsigned types.