Update libxfiles and apps to new new_write_request
This changes the API for `new_write_request` to include a "transaction
type" (feedforward/incremntal/batch) and a number of expected
outputs (only expected to be non-zero for incrmental/batch. This is not
supported by any backend Dana processing at present and the use of this
API will be treated as feedforward requests.