Add block-write enabled SRAMElementIncrement.scala
This includes a new module, SRAMElementIncrement which implements all
features of SRAMElement, but also allows for:
  * block writes overwriting existing data
  * block writes that increment existing data element-wise

These two new types of writes are necessary to enable error
backpropagation and specifically the accumulation of errors before
multiplication with an activation function derivative.

The Register File has been updated to use this new module instead of
SRAMElement.

Block writes are currently untested.