Untested atanh error function
This adds untested support for an atanh error function inside of the
Processing Element module. The error function, a part of the global
configuration, is now written using the modified
`write-fann-config-for-accelerator` tool. This data (really just a bit)
is loaded into the Transaction Table during a global configuration read.

This needs to be integrated with a PE state to apply the atanh function
after computing the error difference.