Fix write count for FEEDFORWARD/LEARN_FEEDFORWARD
The Register File write count was not being properly handled for the two
types of feedforward computation---those without learning and those with
learning. The latter case requires twice the number of writes to the
Register File (as PEs have to write last layer outputs and deltas). The
write count is set by the Control module upon receipt of a response from
the Cache. The Control module, therefore, needs to know that this Cache
response belongs to a transaction with learning and that this
transaction is in its last layer. However, `inLast` is set to late for
this to work.

This defines a new state bit, `inLastEarly`, that asserts as soon as the
last PE in a layer is assigned. This bit, along with the state of the
transaction, is used to determine if the number of expected writes
should be doubled or not.

To help clarify some of the bits related to timing, I've added some WIP
documentation, timing.md.