Fix bug in atanh-lut, now uses correct range
The atanh-lut tool was using a round instead of a ceiling to compute the
number of bits required to represent a number. The original version had
a kludge that would then output values not on the full 31-bit
range (Java is silly and doesn't have unsigned ints, so we have to
respect a +-2**31 range). This fixes this problem so outputs are now on
the full range.