Move state updates to PE Reqs, Fix write count
The TTable state update logic now happens when the Control request is
allowed to leave, i.e., when isPeReq asserts. Additionally, the Register
File write count is now being (properly, though unnecessarily) set to
zero when we're in an ERROR_BACKPROPAGATE state and we hit the first
layer. This works because we really don't need to write back the delta
values to the Register File as these are one-time-use.