Weight Updates Look OK thru Error Backprop Stage
The weight updates for the first layer (which occurs during the error
backpropagation state to save on PE allocations) looks okay. I'm seeing
small updates for values which have a non-zero derivative, i.e., their
original neuron "value" (output of the activation function) is non-zero.

The transaction is not currently jumping out of the Error Backprop
stage, however. As such, the test program dies when the first neuron of
the hidden layer gets assigned as if it were in input layer.