Incremental Learning State Machines Done
This finishes the modifications to the state machines to handle
incremental learning. With a hacked up testbench (not committed with
this), I'm seeing what appears to be the correct behavior for all
neurons except for the first neuron, i.e., I can overtain the network by
repeatedly training on the same input--output vector.

This version does not update the bias.