Fix weight update addr bug
This fixes an addressing bug for the weight update stage where the DW
address (which should be the start of the inputs which need to be read)
was being passed in with a node count offset. This is now fixed.

This also includes commented out code for including a fixed learning
rate.