Add learning rate support to X-Files/Dana
The Cache now understands the format of the learning rate inside of our
reduced FANN configuration and will return this learning rate to the
Transaction Table during a global info request. This required expanding
the port width between the cache and the Transaction Table by 16 (the
current, fragile size of the learning rate).

This includes modifications to the binary configuration generation tool.
The learning rate is now placed in a location in the global info of the
reduced FANN configuration (offset 96, width 16). The documentation has
been updated to reflect this. The learning rate is not currently used
inside of X-FILES/DANA.

This fixes #23.