Add L2 Regularization via Weight Decay
This adds support for constant weight decay based on a parameter,
lambda, defined in the binary NN configuration. This is currently
hardcoded in the binary configuration tool to 0.008 which is the
smallest representable number (put through a floor function) in our
smallest decimal point configuration (7).

This currently seems to work, but increases in lambda can drive the
system into a semi-unstable state. DANA cannot currently recover from a
weight decay that results in an output being driven rail-to-rail (i.e.,
1024->0). Due to the small derivative of a rail, DANA never seems to get
a high enough valued weight update to drive the value back to its
correct rail. Small values of lambda prevent rail-to-rail transitions
and allow the weight update driven by the error to prevent this
transition.

I'm going to close #20 with this, but there are likely still bugs or
oddities (like the rail-to-rail transition) outlined above that need to
be sorted out.