Batch learning looks OK
Fixes #25, fixes #28

This needs more testing from a randomly initialized network. I'm seeing
decreasing MSE when repeatedly training using a batch size of 10. A
batch size of 100 has MSE going up and down before settling close to the
start.

This does not address #29 as the number of epochs are not used to temper
the weight updates (and is possibly the reason why the MSE isn't as good
for a batch size of 100).