Add user function for setting the learning rate
Fixes #29

The learning rate is now set from userland since we need a division to
compute `learning rate / number of batch items`. We can then piggy back
off of the core's multiplier/divider instead of having to burn logic
with one of our own on each PE. This calms the weight updates for larger
batch sizes.