Add fixed variant to fann-train-to-c-header
This adds an explicit tool, "fann-train-to-c-header-fixed", which is a
variant of the existing "fann-train-to-c-header". The new tool expects
all fixed point files (FANN configuration and data file) and will not
infer the decimal point. This is necessary if we're trying to enforce a
specific decimal point for a randomly initialized network (which may
start with a decimal point out of the bounds of what we can represent).