Fix batch learning bias bug, xor stuff, Makefile
This fixes a bug related to batch learning not properly handling bias
updates for batch items that _are not_ the first item.

This includes a new training file for an XOR network as well as a test
program, xor-batch.c, to train this network (unsuccessfully at present)
with X-Files/Dana.

Discussion of batch training failures in #25.