Update FANN build process, add FANN-rv static libs
This updates the way that FANN is built and included in the project. Two
versions of FANN are built:
  * host-machine native (likely x86) shared libraries
  * RISC-V static libraries

Instead of building directly in the FANN submodule, like before, these
two versions are built inside of the build directory. The old symlinks
in usr/lib have been redirected and a new set of symlinks were added in
usr/lib-rv.