Add generic fann-batch RISC-V test program
This adds a new test program, fann-batch, which will run batch training
on an arbitrary NN binary configuration and arbitrary FANN training
file. This paves the way for gathering mass data on the FPGA for
different FANN configurations. This required having static RISC-V FANN
libraries so we have access to `fann_read_train_from_file`.

FANN does not like to build with the generic ELF/Newlib toolchain but
wants Linux-ELF/glibc. This increases binary size substantially, but
does not appear to detrimentally affect any of the other binaries.
Consequently, the Makefile now uses the Linux/glibc toolchain for all
binaries.

The original location of the libxfiles symlink has been moved to lib-rv
to differentiate what is a host library vs. a RISC-V library.