Add RISC-V cycle counter to fann-batch
Using the available RISC-V performance counters, specifically the cycle
counter, I'm now able to get an accurate cycle count when running batch
training on an X-Files/Dana enabled Rocket Chip. This, coupled with a
new function (internal to fann-batch) that reads a binary NN
configuration to get the number of connections, enables the reporting of
Connection Updates per Cycle (CUPC) with a new command line option. To
enable accurate performance counting, there is now a "performance"
option which uses different internal code that avoids any comparisons
and printfs and, instead, runs for a fixed number of epochs.