Add explicit TTable location bit, bug fixes
The Register File location bit cannot be defined as the LSB of the
current layer number. This is because jumps from non-adjacent layers are
possible (specifically when doing batch learning and jumping form the
last layer to the first to do weight updates). This causes a bug where
the location bit is the same for the last layer and for the first if
defined as the LSB of the current layer number.

This adds an assertion that checks for what would be an overwrite of the
total number of writes to the Register File by the control module. To
avoid this issue, the location is now determined by the Transaction
Table. The location bit is initialized to zero and then toggled whenever
that transaction starts any work on a new layer.

This also fixes a state transition bug in the PE where the transition
out of e_PE_DELTA_WRITE_BACK was incorrect.