Fix MSE reporting for symmetric AFs in fann-train
FANN treats MSE strangely for sigmoid symmetric activation functions in
the output layer. To try and make symmetric and non-symmetric activation
functions report similar MSEs, FANN divides the error by two inside
fann_update_MSE. This causes the overall MSE to be 1/4 of what it should
be my any normal math computation.

This fixes that by multiplying the FANN-computed MSE by four if we see a
symmetric activation function in the output layer.