Add fann-random seed (disable auto-seed in FANN)
This adds a `-s` option to fann-random to set the seed used when
generating a network. To get this to work, FANN needs to be built with
FANN_NO_SEED defined. The top-level xfiles-dana Makefile now supplies
this when building both host and RISC-V versions of the FANN library.
The top-level Makefile relies on bash's $RANDOM to supply the seed when
generating networks with fann-random now.