Fix a bug related to rate of ANTW memory access
This limits the rate that memory accesses are allowed to exit the ANTW
to one every two cycles. I was seeing some issues related to dropped
requests (which I can't figure out is my problem or a rocket/uncore
bug).

This also includes an assertion which will mostly catch the case where a
packet gets dropped---if the config reorder buffer (unit that packs XLen
packets into block-sized chunks) tries to overwrite a valid entry this
will fire. This generally seems to avoid the alternative which is a
hang. Here, the ANTW never gets all the expected packets back so it hits
the end of what it gets and just sits there.